pyproject.toml
[tool.poetry]
name = "support-tg-bot"
version = "0.1.0"
description = ""
authors = ["pelkoa-glitch <hvngmvnvibe@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django-environ = "^0.12.0"
python-telegram-bot = "^22.0"
pydantic = "^2.11.3"
pydantic-settings = "^2.8.1"
dishka = "^1.5.2"
faststream = {extras = ["cli", "kafka"], version = "^0.5.39"}
watchfiles = "^1.0.5"
aiosqlite = "^0.21.0"


[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
